pyproject.toml
[tool.uv.sources]
torch = [
  { index = "pytorch-cpu", marker = "platform_system != 'Darwin'"},
]
torchvision = [
  { index = "pytorch-cpu", marker = "platform_system != 'Darwin'"},
]
torchaudio = [
  { index = "pytorch-cpu", marker = "platform_system != 'Darwin'"},
]

[project]
name = "aws-hackathon"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "beautifulsoup4>=4.12.2,<5.0.0",
    "google-search-results>=2.4.1,<3.0.0",
    "google-api-python-client>=2.130.0,<3.0.0",
    "huggingface-hub[inference]>=0.23.2,<1.0.0",
    "networkx>=3.1,<4.0.0",
    "fake-useragent>=1.5.0,<2.0.0",
    "pyarrow>=14.0.0,<18.0.0",
    "wikipedia>=1.4.0,<2.0.0",
    "qdrant-client~=1.9.2",
    "weaviate-client>=4.8,<5.0.0",
    "cohere>=5.5.3,<6.0.0",
    "faiss-cpu>=1.8.0,<2.0.0",
    "types-cachetools>=5.3.0.5,<6.0.0",
    "pymongo>=4.6.0,<5.0.0",
    "supabase~=2.6.0",
    "certifi>=2023.11.17,<2025.0.0",
    "fastavro>=1.8.0,<2.0.0",
    "redis>=5.0.1,<6.0.0",
    "metaphor-python>=0.1.11,<1.0.0",
    'pywin32>=306,<307; sys_platform == "win32"',
    "langfuse~=2.53.1",
    "metal-sdk>=2.5.0,<3.0.0",
    "markupsafe>=2.1.3,<4.0.0",
    "boto3~=1.34.162",
    "numexpr>=2.8.6,<3.0.0",
    "qianfan==0.3.5",
    "pgvector>=0.2.3,<1.0.0",
    "pyautogen>=0.2.0,<1.0.0",
    "langchain~=0.3.3",
    "elasticsearch>=8.12.0,<9.0.0",
    "pytube>=15.0.0,<16.0.0",
    "dspy-ai>=2.4.0,<3.0.0",
    "assemblyai>=0.33.0,<1.0.0",
    "litellm>=1.44.0,<2.0.0",
    "chromadb>=0.4,<1.0.0",
    "zep-python>=2.0.0,<3.0.0",
    "langchain-groq>=0.1.9,<1.0.0",
    "langchain-pinecone>=0.1.3,<1.0.0",
    "youtube-transcript-api>=0.6.2,<1.0.0",
    "markdown>=3.7,<4.0.0",
    "upstash-vector>=0.5.0,<1.0.0",
    "gitpython>=3.1.43,<4.0.0",
    "kubernetes>=30.1.0,<32.0.0",
    "json-repair>=0.25.2,<1.0.0",
    "langchain-google-calendar-tools>=0.0.1,<1.0.0",
    "langchain-milvus>=0.1.1,<1.0.0",
    "langwatch==0.1.16",
    "langsmith~=0.1.137",
    "yfinance>=0.2.40,<1.0.0",
    "wolframalpha>=5.1.3,<6.0.0",
    "astra-assistants[tools]~=2.2.6",
    "composio-langchain==0.5.42",
    "spider-client>=0.0.27,<1.0.0",
    "nltk>=3.9.1,<4.0.0",
    "lark>=1.2.2,<2.0.0",
    "jq>=1.8.0,<2.0.0",
    "pydantic-settings==2.4.0",
    'torch==2.4.1+cpu; platform_system != "Darwin"',
    'torchvision==0.19.1+cpu; platform_system != "Darwin"',
    'torch==2.4.1; platform_system == "Darwin"',
    'torchvision==0.19.1; platform_system == "Darwin"',
    "pdfplumber>=0.5.28",
    "pymupdf>=1.24.11",
    "ragstack-ai-knowledge-store>=0.2.1,<1.0.0",
    "duckduckgo-search>=6.3.4,<7.0.0",
    "opensearch-py>=2.7.1,<3.0.0",
    "langchain-google-genai>=2.0.1,<3.0.0",
    "langchain-cohere>=0.1.5,<1.0.0",
    "langchain-anthropic>=0.1.23,<1.0.0",
    "langchain-astradb~=0.5.2",
    "langchain-openai~=0.2.2",
    "langchain-google-vertexai~=2.0.5",
    "langchain-groq>=0.1.9,<1.0.0",
    "langchain-pinecone~=0.2.0",
    "langchain-mistralai~=0.2.0",
    "langchain-chroma>=0.1.1,<1.0.0",
    "langchain-aws>=0.1.16,<1.0.0",
    "langchain-mongodb>=0.1.6,<1.0.0",
    "langchain-nvidia-ai-endpoints~=0.3.0",
    "langchain-google-community~=2.0.1",
    "langchain-elasticsearch>=0.2.0,<1.0.0",
    "langchain-ollama>=0.2.0,<1.0.0",
    "sqlalchemy[aiosqlite,postgresql_psycopg2binary,postgresql_psycopgbinary]>=2.0.36,<3.0.0",
    "atlassian-python-api>=3.41.16,<4.0.0",
    "mem0ai>=0.1.26,<1.0.0",
    "pdf2docx==0.5.6",
    "pytesseract==0.3.13",
    "camelot-py==0.11.0",
    "fuzzywuzzy==0.18.0",
    "scikit-learn>=1.5.2",
    "chromadbx>=0.0.8",
    "rerankers>=0.8.0",
    "azure-ai-documentintelligence==1.0.2",
    "s3fs>=0.4.2",
    "st-files-connection>=0.1.0",
    "streamlit>=1.48.1",
    "pdf2image>=1.17.0",
]

[tool.uv]
dev-dependencies = [
    "pytest-instafail>=0.5.0",
    "types-redis>=4.6.0.5",
    "ipykernel>=6.29.0",
    "mypy>=1.11.0",
    "ruff>=0.6.2,<0.7.0",
    "httpx>=0.27.0",
    "pytest>=8.2.0",
    "types-requests>=2.32.0",
    "requests>=2.32.0",
    "pytest-cov>=5.0.0",
    "pandas-stubs>=2.1.4.231227",
    "types-pillow>=10.2.0.20240213",
    "types-pyyaml>=6.0.12.8",
    "types-python-jose>=3.3.4.8",
    "types-passlib>=1.7.7.13",
    "pytest-mock>=3.14.0",
    "pytest-xdist>=3.6.0",
    "types-pywin32>=306.0.0.4",
    "types-google-cloud-ndb>=2.2.0.0",
    "pytest-sugar>=1.0.0",
    "respx>=0.21.1",
    "pytest-asyncio>=0.23.0",
    "pytest-profiling>=1.7.0",
    "pre-commit>=3.7.0",
    "vulture>=2.11",
    "dictdiffer>=0.9.0",
    "pytest-split>=0.9.0",
    "pytest-flakefinder>=1.1.0",
    "types-markdown>=3.7.0.20240822",
    "packaging>=24.1,<25.0",
    "asgi-lifespan>=2.1.0",
    "pytest-github-actions-annotate-failures>=0.2.0",
    "pytest-codspeed>=3.0.0",
    "blockbuster>=1.1.1,<1.2",
]

[[tool.uv.index]]
name = "pytorch-cpu"
url = "https://download.pytorch.org/whl/cpu"
explicit = true
